# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/public/traps/summary": {
      "get": {
        "operationId": "TrapsController_summary",
        "summary": "Ringkasan deteksi hama",
        "description": "Total event per jenis, deteksi hama hari ini, jumlah trap. Bisa filter ?deviceId=",
        "parameters": [
          {
            "name": "deviceId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public/traps"
        ]
      }
    },
    "/v1/public/traps/{deviceId}/latest": {
      "get": {
        "operationId": "TrapsController_latestOne",
        "summary": "Event terkini satu trap",
        "parameters": [
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public/traps"
        ]
      }
    },
    "/v1/public/traps/{deviceId}/history": {
      "get": {
        "operationId": "TrapsController_history",
        "summary": "Riwayat event trap",
        "description": "Untuk grafik tren. Filter ?eventType=pest_detected dan ?limit=",
        "parameters": [
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eventType",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public/traps"
        ]
      }
    },
...
    "/v1/cms/schedule": {
      "get": {
        "operationId": "ScheduleCmsController_listAll",
        "summary": "Semua jadwal tanam (admin)",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/schedule"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "post": {
        "operationId": "ScheduleCmsController_create",
        "summary": "Tambah jadwal tanam",
        "description": "Estimasi panen otomatis jika tidak diisi (default umur padi 115 hari).",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cms/schedule"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/v1/cms/schedule/upcoming-harvest": {
      "get": {
        "operationId": "ScheduleCmsController_upcoming",
        "summary": "Jadwal mendekati panen",
        "description": "Default 7 hari ke depan. ?days=",
        "parameters": [
          {
            "name": "days",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/schedule"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/v1/cms/schedule/{id}": {
      "get": {
        "operationId": "ScheduleCmsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/schedule"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "put": {
        "operationId": "ScheduleCmsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/schedule"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "ScheduleCmsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/schedule"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "NAWASENA Center API",
    "description": "Backend NAWASENA Center untuk PPK Ormawa RDC UAD.\n\n## Cara pakai\n- Endpoint **public/** bisa diakses tanpa login (untuk website publik).\n- Endpoint **cms/** butuh login admin. Klik tombol **Authorize** di kanan atas, masukkan token (format: `Bearer <token>` tidak perlu, cukup token-nya).\n\n## Format response\nSemua response: `{ success: true, data: {...} }` atau `{ success: false, error: {...} }`.\n\n## Autentikasi\nLogin via Supabase Auth untuk dapat access_token, lalu Authorize di sini.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Masukkan access_token dari Supabase Auth (tanpa kata Bearer)"
      }
    },
    "schemas": {
...
      "CreateScheduleDto": {
        "type": "object",
        "properties": {
          "farmerName": {
            "type": "string",
            "example": "Pak Sutrisno",
            "description": "Nama petani"
          },
          "fieldName": {
            "type": "string",
            "example": "Sawah Blok A",
            "description": "Nama/kode lahan"
          },
          "padukuhan": {
            "type": "string",
            "example": "Karangtengah",
            "description": "Padukuhan (untuk tanam serentak per wilayah)"
          },
          "plantDate": {
            "type": "string",
            "example": "2026-06-20",
            "description": "Tanggal tanam (YYYY-MM-DD)"
          },
          "estimatedHarvest": {
            "type": "string",
            "example": "2026-10-13",
            "description": "Perkiraan panen. Kosongkan untuk hitung otomatis dari umur padi."
          },
          "cropDays": {
            "type": "number",
            "example": 115,
            "description": "Umur padi (hari) untuk hitung estimasi panen otomatis. Default 115."
          },
          "notes": {
            "type": "string",
            "example": "Varietas IR64"
          }
        },
        "required": [
          "farmerName",
          "plantDate"
        ]
      },
      "UpdateScheduleDto": {
        "type": "object",
        "properties": {
          "farmerName": {
            "type": "string",
            "example": "Pak Sutrisno",
            "description": "Nama petani"
          },
          "fieldName": {
            "type": "string",
            "example": "Sawah Blok A",
            "description": "Nama/kode lahan"
          },
          "padukuhan": {
            "type": "string",
            "example": "Karangtengah",
            "description": "Padukuhan (untuk tanam serentak per wilayah)"
          },
          "plantDate": {
            "type": "string",
            "example": "2026-06-20",
            "description": "Tanggal tanam (YYYY-MM-DD)"
          },
          "estimatedHarvest": {
            "type": "string",
            "example": "2026-10-13",
            "description": "Perkiraan panen. Kosongkan untuk hitung otomatis dari umur padi."
          },
          "cropDays": {
            "type": "number",
            "example": 115,
            "description": "Umur padi (hari) untuk hitung estimasi panen otomatis. Default 115."
          },
          "notes": {
            "type": "string",
            "example": "Varietas IR64"
          },
          "status": {
            "type": "string",
            "example": "planted",
            "description": "planned | planted | harvested"
          }
        }
      },
...
    }
  }
}